+++ lib/types/errors.go
@@ -64,6 +64,7 @@ var (
 	ErrCacheNotFound     = errors.New("cache not found")
 	ErrConditionNotFound = errors.New("condition not found")
 	ErrRateLimitNotFound = errors.New("rate limit not found")
+	ErrPluginNotFound    = errors.New("plugin not found")
 	ErrKeyAlreadyExists  = errors.New("key already exists")
 	ErrKeyNotFound       = errors.New("key does not exist")
 	ErrPipeNotFound      = errors.New("pipe was not found")